+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/DockerTestKafkaEventSource.java
@@ -24,7 +24,6 @@ import io.telicent.smart.cache.sources.Event;
 import io.telicent.smart.cache.sources.EventSourceException;
 import io.telicent.smart.cache.sources.Header;
 import io.telicent.smart.cache.sources.kafka.policies.KafkaReadPolicies;
-import io.telicent.smart.cache.sources.kafka.serializers.RdfPayloadDeserializer;
 import io.telicent.smart.cache.sources.kafka.sinks.KafkaSink;
 import io.telicent.smart.cache.sources.memory.SimpleEvent;
 import io.telicent.smart.cache.sources.offsets.MemoryOffsetStore;
@@ -491,27 +490,6 @@ public class DockerTestKafkaEventSource {
                                     .stream()
                                     .map(LoggingEvent::getFormattedMessage)
                                     .anyMatch(m -> CS.contains(m, "Kafka reported error deserializing")));
-    }
-
-    @Test
-    public void givenRdfRecordWithIncorrectContentTypeHeader_whenPollingKafkaForPayloadWithEagerParsing_thenRecordDeserializationErrorOccurs() {
-        // Given
-        injectRdfEventWithWrongContentType();
-
-        // When
-        TestLogger testLogger = TestLoggerFactory.getTestLogger(KafkaEventSource.class);
-        testLogger.clearAll();
-        KafkaRdfPayloadSource<Bytes> source = KafkaRdfPayloadSource.<Bytes>createRdfPayload()
-                                                                   .bootstrapServers(this.kafka.getBootstrapServers())
-                                                                   .topic(KafkaTestCluster.DEFAULT_TOPIC)
-                                                                   .consumerGroup("record-deserialization-02")
-                                                                   .consumerConfig(
-                                                                           RdfPayloadDeserializer.EAGER_PARSING_CONFIG_KEY,
-                                                                           "true")
-                                                                   .keyDeserializer(BytesDeserializer.class)
-                                                                   .build();
-
-        verifyDeserializationErrorOnPoll(source, testLogger);
     }
 
     @Test(expectedExceptions = RdfPayloadException.class)