+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/tombstones/LazyData.java
@@ -13,19 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.telicent.smart.cache.sources.kafka.serializers;
+package io.telicent.smart.cache.sources.kafka.tombstones;
 
-import java.util.Map;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.telicent.smart.cache.payloads.LazyJacksonPayload;
 
-public class TestPayloadDeserializerBadConfig extends TestPayloadDeserializer {
+public class LazyData extends LazyJacksonPayload<Data> {
+    LazyData(ObjectMapper mapper, Class<Data> cls, byte[] rawData) {
+        super(mapper, cls, rawData);
+    }
 
-    private final Map<String, ?> BAD_PARSING_CONFIGURATION =
-            Map.of(RdfPayloadDeserializer.EAGER_PARSING_CONFIG_KEY, 1234);
-
-    @Override
-    protected RdfPayloadDeserializer createPayloadDeserializer() {
-        RdfPayloadDeserializer deserializer = super.createPayloadDeserializer();
-        deserializer.configure(BAD_PARSING_CONFIGURATION, false);
-        return deserializer;
+    public LazyData(Data value) {
+        super(value);
     }
 }